All files for website
I had to take them out of my folder in my local file and upload, so the sources that I call in the HTML might cause issues. I will send you a zip file of all the entire project file so that you can decompress and run the html without issue. Sorry for turning it in late! I hope you're thoroughly impressed with the amount of work I put into creating this.